# tools/schematics/todo.openapi.json
{"openapi":"3.0.0","info":{"title":"Todo API","description":"An api for a simple todo application.","version":"1.0","contact":{}},"tags":[{"name":"todo","description":""}],"servers":[],"components":{"securitySchemes":{"oauth2":{"type":"oauth2","flows":{"implicit":{"scopes":{},"authorizationUrl":"http://localhost:3000/api/auth/token"}}}},"schemas":{"CreateTodo":{"type":"object","properties":{"value":{"type":"string","description":"The todo","example":"Cook Meal"},"userId":{"type":"string","description":"The user that created / owns the todo.","example":""},"priority":{"type":"string","default":"medium","enum":["low","medium","high"],"example":"high"},"description":{"type":"string","description":"A longer description of the todo","example":"Cook a very delicious bowl of spaghetti bolognese."},"status":{"type":"string","default":"open","description":"The status of the todo.","example":"done","enum":["open","done"]},"deadline":{"type":"number","description":"The deadline of the todo.","example":""}},"required":["value"]},"Todo":{"type":"object","properties":{"id":{"type":"string","example":"febad8f8-6dc4-4493-a291-5630b6d54ffb"},"value":{"type":"string","description":"The todo","example":"Cook Meal"},"priority":{"type":"string","enum":["low","medium","high"],"example":"high"},"description":{"type":"string","description":"A longer description of the todo","example":"Cook a very delicious bowl of spaghetti bolognese."},"status":{"type":"string","description":"The status of the todo","example":"done","enum":["open","done"]},"deadline":{"type":"number","description":"The deadline of the todo","example":""},"userId":{"type":"string","description":"The user that created / owns the todo.","example":""}},"required":["id","value","priority","description","status","deadline"]},"UpdateTodo":{"type":"object","properties":{"value":{"type":"string","description":"The todo","example":"Cook Meal"},"priority":{"type":"string","enum":["low","medium","high"],"example":"high"},"description":{"type":"string","description":"A longer description of the todo","example":"Cook a very delicious bowl of spaghetti bolognese."},"status":{"type":"string","description":"The status of the todo","example":"done","enum":["open","done"]},"deadline":{"type":"number","description":"The deadline of the todo","example":""},"userId":{"type":"string","description":"The user that created / owns the todo.","example":""}}}}},"paths":{"/api/todo":{"post":{"operationId":"create","summary":"","description":"Create a new todo.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTodo"}}}},"responses":{"201":{"description":"Successfully created the todo.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Todo"}}}}},"tags":["todo"],"security":[{"oauth2":[]}]},"delete":{"operationId":"deleteByFilter","summary":"","description":"Delete todos by a filter.","parameters":[{"name":"status","required":true,"in":"query","example":"done","schema":{"enum":["open","done"],"type":"string"}}],"responses":{"200":{"description":"Successfully deleted the todos matching the filter.","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["todo"],"security":[{"oauth2":[]}]},"get":{"operationId":"getAll","summary":"","description":"Get all your todos.","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Todo"}}}}}},"tags":["todo"],"security":[{"oauth2":[]}]}},"/api/todo/all-users":{"get":{"operationId":"getForAllUsers","summary":"","description":"Get all todos for all users.","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Todo"}}}}}},"tags":["todo"],"security":[{"oauth2":["todo:read:all"]}]}},"/api/todo/{id}":{"put":{"operationId":"update","summary":"","description":"Update one of your todos.","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTodo"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Todo"}}}}},"tags":["todo"],"security":[{"oauth2":[]}]}}}}